Add argOrdinal field to descriptor map
Also change how the descriptor map is derived.  When clustering
POD kernel arguments, attach a "kernal_arg_map" metadata node
to the kernel function to fully describe the mapping.  This could
be more robust, in the face of potential inlining that could occur
later.

When that metadata is present, it has full information on the kernel
argument argument mapping.  Otherwise, we just derive the descriptor map
entries directly from the kernel arguments.