Fix constants generation in presence of <4 x i8> constant
In the loop generating all SPIR-V constants, there was an
incorrect early function return in the case of a <4 x i8>
constant.  The symptom is that a later constant, such as
for an undef aggregate value, was not generated.  That combined
with other bookkeeping accidents to cause an OpCompositeInsert
to have the incorrect initial composite value ID.

Partial fix for issue 14