Use LLVM's AtomicRMWInst for most of the atomics instructions.
LLVM provides an AtomicRMWInst that satisfies most of the spirv.atomic
'fake' intrinsics we used previously, so lets switch over to use that.
This also allows other languages that aren't coming from a CL
perspective to use AtomicRMWInst and have clspv output the correct
instructions.

I also fixed a bug with the mem_fence replacements - having two maps of
const char* and then comparing if the two pointers matched would compare
the two separate binary blobs of the const char* (EG. their pointers)
which might not match (depending on the linker). Instead I changed it to
a std::map where the second element is a std::tuple.