Zero-initialize all stack variables
This helps avoid cases where we get a ConstantExpr of fadd over
undef.  The SPIR-V code generation pass can't handle that kind of
thing.
This is a bit of a hack.

Also generalize the pointer bitcast replacement to handle source
pointer types that point to arrays that are larger than the destination
pointer element type.