Fix codegen for logical operations on bool vectors
Bitwise instructions were being emitted. The tests added
only cover And and Or as I couldn't find a way of
producing Xor on boolean vectors which seems to line up
with a few observations from the OpenCL specification:

- there is no xor logical operator (although ^^ has been
  reserved for that use)
- boolean vectors aren't explicitly specified (even
  though booln (2,3,4,8 and 16) is a reserved type)
- the result of relational operators applied to two
  vectors is guaranteed to be a vector of signed integers
  (this is the case that made me discover the issue)

Clang's omission of the sign extension to int vectors
when the result of multiple relational operators is
only used by a logical operation is an optimisation.
Since there is no logical xor operator it is
reasonable for Clang to always sign extend the result
of relational operators (according to §6.3d of the
OpenCL 1.2 specification) that feed into an xor
operator and perform the operation as bitwise.

An optimisation similar to that done for And and Or
could be made by clspv in cases where the results of
relational operators are only used by xor operators
but should IMHO be discussed separately from this
change.

Also tidy up a few related things:

- use isIntOrIntVectorTy(1) where it makes the code clearer
- tighten the check in SPIRVProducerPass::GenerateInstruction
  that special cases Instruction::Xor for booleans. It was
  allowing OpLogicalNot to be emitted in all cases where one
  of the operands was constant, which is correct only when the
  constrant is 1. The value of the constant is now checked.

Signed-off-by: Kévin Petit <kpet@free.fr>