Add support for the upsample built-in function (#271)
The tests added here are a compromise between succinctness,
ease of maintenance and coverage. The second argument to upsample
is a constant and its conversion to the result type is folded
into the constant. However, using two non-constant arguments
leads to a much longer and brittler test if one wants to ensure
that their relative ordering is correct.

Signed-off-by: Kévin Petit <kpet@free.fr>