Add support for changing the location of opt and passes.
This adds some command-line arguments to clsvp-opt to re-locate both
LLVM opt and the clsvp pass library.

Given that the tool now needs to handle its own arguments and pass
anything it doesn't understand down to 'opt', before parsing the
command-line, it splits the argument space in two lists: one for itself
and another one for 'opt'.