Introduce a mechanism to call specific SPIR-V instructions
Directly calling a specific SPIR-V instruction is a common
idiom which should be shorter to write than it currently is.
Besides, having a centralised scheme that deals with encoding
SPIR-V intrinsics removes the need to worry about issues like

This change introduces a utility function that inserts a call
to a SPIR-V instruction passed as a parameter. It automatically
inserts required functions into the module if needed and
guarantees they have a unique name with a simple mangling
scheme.

The current encoding makes the IR a bit less readable as
the functions used to encode SPIR-V instructions are named
after the opcode number rather than the instruction name.
This can be improved upon later if felt necessary.

This change also uses the new mechanism for the mul_hi/mad_hi
built-in functions, fixing the issue described in #308 for
these.

Signed-off-by: Kévin Petit <kpet@free.fr>