Add a tool to run individual clspv passes. (#283)
Add clspv-opt: a tool to run individual clspv passes.

This fixes #263.  It adds a wrapper around LLVM's opt that allows
running any of the clspv passes on an LLVM IR module.

In this initial implementation, the developer must:

1- Generate IR using the `-print-before=` or `-print-after=` options to
   clspv.
2- Run clspv-opt on the generated LLVM IR file.

This required some re-factoring of `clspv_core`:

1- All the passes are now in a separate shared library (`clspv_passes`).
   This is required by `opt`.
2- The core library continues to be static.

Given that clspv-opt needs to handle its own arguments and pass
anything it doesn't understand down to 'opt', before parsing the
command-line, it splits the argument space in two lists: one for itself
and another one for 'opt'.

Currently, clspv-opt is only supported on Linux and MacOS (#307).  Support for windows will require several changes to it.