[RFC] Experimental C++ support
This change adds experimental support for C++ via Clang's
OpenCL C++ mode.

In that mode Clang uses OpenCL's generic address space for
a number of things so I've just added it to
clspv::AddressSpace to align clspv's internal convention
with what Clang does a bit better.

This currently has to be used with -inline-entry-points to
maximise the chances that all traces of generic address
space will have been removed before passes that don't
support it. I have another change that plugs LLVM's address
space inference pass and enables quite a bit more code to be
compiled (and the use of the generic address space generally)
but it requires a couple of minor changes to LLVM.

Kernel function overloads are forbidden in the FrontendPlugin
with a custom diagnostic message. Kernel function names
are demangled before the descriptor map creation and the
absence of overload guarantees name unicity.

Signed-off-by: Kévin Petit <kpet@free.fr>