[C++] Replace generic address space before SPIR-V production
This enables support for the generic address space in all
cases where the final address space can be inferred before
the SPIR-V production.

This should provide a step change in the amount of C++ code
that can be compiled.

Nota: both the Clang and LLVM updates are required to make
this work.

Fixes #357.

Signed-off-by: Kévin Petit <kpet@free.fr>