Add flag to emit LLVM IR after parsing.
This adds the flag `--emit-ir` which can be used in conjunction with
`clspv-opt` to develop and debug clspv passes.

The flag emits LLVM IR right after the module has been parsed.  It
seemed the most natural place to dump the IR, because it gives the most
flexibility.  Once the initial IR has been generated, all the passes up
to the problematic one can be applied via `clspv-opt`.