Prevent OpVariable from being produced with an OpUndef initializer
This was avoided most of the time by luck because the OpUndef
wouldn't be in the value map. This change makes it a guarantee
and the added assert will catch any cases where initializers
are omitted by accident.

The new test case has been produced by reducing a kernel from
the SHOC benchmark suite.

Signed-off-by: Kévin Petit <kpet@free.fr>