Initial support for non-uniform NDRanges (#572)
This implementation uses the approach outlined in #522 and push constants
for all the values that need to be communicated by the application/runtime.
The push constant space required is quite large in the worst case (which is
unlikely to be a common occurrence in typical code). There are definitely
applications where using specialisation constants everywhere would result
in a ridiculous amount of recompilation and likely others where saving some
push constant space would be a win (especially once POD arguments will
use push constants).

The CTS tests require that reqd_work_group_size be treated not as a hard
requirement but more as an upper bound/hint which seems odd. I've created
https://github.com/KhronosGroup/OpenCL-Docs/issues/267 to discuss this.
This change satisfies the CTS and is good enough to pass the entire
non_uniform_work_group CTS suite with the appropriate clvk changes.

Fixes #522.

Signed-off-by: Kévin Petit <kpet@free.fr>