Add stubs for lowering long-vector types
Add a new, experimental command line option, --long-vector, to turn on
support for lowering long-vector types which is implemented by
LongVectorLoweringPass. For now, this pass does nothing; subsequent
commits will gradually bring in new features.

When this option is provided, accept long-vector types except when used
as kernel parameters. This feature is intentionally left unsupported for
now and should be revisited when there is definite proof that kernel
parameters can be safely lowered as well.

Add tests covering the frontend changes.

Introduce CustomDiagnosticUnsupportedKernelArgs. Remove explicit
diagnostic ID and rely on C++ to generate appropriate values for the
CustomDiagnosticType enumeration.

Update options.ll test to ensure it can be used with clspv-opt.

Signed-off-by: Marco Antognini <marco.antognini@arm.com>