Recursively lower instructions
Lower InsertElementInst, ExtractElementInst, ShuffleVectorInst,
UndefValue, ConstantAggregateZero, ConstantDataVector and pointer types.

With support for these instructions, it is now possible to use recursion
to implement support for BinaryOperator with the existing tests.

Strengthen checks in getEquivalentTypeImpl. Now that this function is
recursive it cannot be const anymore.

Enable the disabled llvm_unreachable statements to catch bugs now that
enough functionalities are implemented.

Add test to cover the newly supported constants and instructions.

Signed-off-by: Marco Antognini <marco.antognini@arm.com>