Add logic to lower OpenCL builtin functions
Initiate lowering of OpenCL builtin functions with the overloads of max
for int8, int16, float8 and float16 types.

The implementation is currently based on a very simple mapping of
mangled names, from the vector overload to its scalar counterpart. This
solution is known to have limitations, especially w.r.t. scaling to
support all OpenCL builtin functions, yet it is trivial and seemed
sufficient for this prototype. It should nonetheless be replaced with a
generic solution later on.

Add two new test cases, covering the SPIR-V generation. Extend an
existing test case, covering the case where multiple overloads of the
same builtin function are used in the same program.

Signed-off-by: Marco Antognini <marco.antognini@arm.com>