Lower BinaryOperator instructions on long-vector (#642)
* Lower BinaryOperator

Introduce some logic to lower instructions that use long-vector
operands. Initiate support with BinaryOperators. Add a test case testing
this lowering.

Lower vector of 8 or more scalars to aggregate of the same number of
scalar elements.

Later, a recursive solution will be implemented. In the meantime, to
make commits smaller and still produce valid LLVM IR, the operands of
the BinaryOperators are converted to equivalent aggregate values.
Similarly, the aggregates generated by the scalarisation of the
BinaryOperators are converted back to vector values.

Signed-off-by: Marco Antognini <marco.antognini@arm.com>

* Add cache for value and type lowering

Reuse previous lowering instead of redoing the same work. The
instruction cache will also be used to remove dead instructions.

Signed-off-by: Marco Antognini <marco.antognini@arm.com>

* Remove dead instructions after lowering

Removal is done on best-effort basis. InstCombine and other optimisation
pass are expected to further simplify the IR generated by the lowering
phase, especially once more instructions are supported.

Update test cases to cover the removal.

Signed-off-by: Marco Antognini <marco.antognini@arm.com>