Add options for fp16 and fp64 and disable unsupported CL3 features (#722)
* Add options for fp16 and fp64 and disable unsupported CL3 features

Some OpenCL applications (e.g. clpeak) conditionally enable FP16/FP64 in
kernels based on the presence of the cl_khr_fp{16,64} macros. When using
the SPIR target Clang always enables all OpenCL extensions and features
potantially resulting in applications attempting to use features not
supported on the target device.

Since Clang always enables all features when using the SPIR target,
it means that clspv must either pass a list of features to be disabled
to Clang (not sure this would be accepted upstream) or disable features
after the target-specific defaults have been applied. The latter is the
approach proposed in this change: extensions and features are disabled
after the creation of the TargetInfo and before the CompilerInstance
is configured to use it.

The new options are enabled by default for backwards compatibility.

Signed-off-by: Kévin Petit <kevin.petit@arm.com>