Allow function private storage pointers
The SPIRV producer has special cases for GEP handling, particularly when
the first index is 0.  However, when using pointer variables incremented
in a for-loop, the GEP operand becomes a PHINode, where the first offset
is not 0.  This causes the associated code in the SPIRV generator to
crash when it does not expect this access chain type.

Signed-off-by: Pedro Olsen Ferreira <pedro.olsenferreira@arm.com>