Support some pointer comparisons when targeting SPIR-V 1.4+
Using OpPtrEqual, OpPtrNonEqual and OpPtrDiff.

No attempt was made to validate that the operands to OpPtrDiff are
"pointers to element numbers in [0, L] in the same array". This means
users can easily produce SPIR-V triggering undefined behaviour when
this doesn't hold.

We could maybe check that both pointers are derived from the same resource
var and even check that all but one indices match when pointers are
produced with a single GEP from a resource var.

Fixes #383

Signed-off-by: Kévin Petit <kevin.petit@arm.com>