ReplacePointerBitcastPass: consider struct of same element as array
The long vector lowering pass lower long vector into structure of the
same element.
In some case, it can produce the following pattern, which is not valid:
%x = OpAccessChain
%y = OpBitcast %type %x
%z = OpLoad %type %y

For arrays, this is prevented by the replace pointer bitcast pass.
We can do the same for the structures generated by the long vector
lowering pass by considering it as a array.