ReplacePointerBitCast: manage indices for gep on multi-dim arrays
When a bitcast of an array of array followed by a gep is detected,
generated a new gep that consider each dimension of the array to
compute the proper indices.

If the source is from the private memory, force the first indice to be
zero (meaning no out-of-bounds possible), as it would be against the
expected spirv behavior later.