longvector: Uses Array instead of Struct
To support long vectors, clspv use to transform them into struct of
the element type of the vector. But it is not ideal because llvm does
not know that all the elements of the structure are the same.
Let's use array instead of struct to help llvm.

No regression on OpenCL CTS + vload_local is now passing

Fixes #780