enable support of non uniform workgroup size by default
Support of non uniform workgroup size also allow to split a NDRange in
several smaller regions.
We need that feature available by default to run the CTS on small
embedded GPUs even in CL1.2.
Allow to disable to feature with '-uniform-group-size'.

This is related to the work in kpet/clvk#367