OpMulExtended: workaround Intel GPU issue (#828)
* OpMulExtended: workaround Intel GPU issue

OpMulExtended only works on Intel GPU with 32bit types.
Workaround it with a new option 'hack-mul-extended' which prevents
clspv from using OpMulExtended.

Also change the SRem implementation not to use OpMulExtended all the
time.

* transform xor and select into abs

Depending on the version of llvm. llvm can do the transformation by
itself.
do not rely on llvm do to it. let's generate it directly

* do not use mul64 for 32x32 in 64