Handle llvm.{u,s}{add,sub}_sat intrinsics (#852)
* Fix add/sub sat test generators

They were still using the old-style pass-selecting options and hadn't
been updated to use the new pass manager. This makes the generators
consistent with the tests.

* Handle llvm.{u,s}{add,sub}_sat

Change handling of the CL builtin functions {add,sub}_sat to generate
the LLVM intrinsics, then lower the intrinsics using the existing code
for {add,sub}_sat.

* Add replace-llvm-intrinsics pass in _sat tests

These OpenCL builtins are now replaced in 2 stages, so both passes need
to be run in order to generate the expected results.