add support for opencl compliant division
This is an implementation based on OpFDiv to avoid loosing too much
performance.
When possible it can be reduced to a single OpFDiv if we are divided
by a constant value.

This implementation passes the CTS for divide/half_divide on Intel,
AMD, ARM and Nvidia.

Also remove kHalfRecip and kHalfDivide from ReplaceOpenCLBuiltinPass
as they are in fact managed through clpsv--.bc.

Apply NoContraction Decoration to a bit more fp operands as GPU like
Intel are doing optimization with some of them (CTS contractions test
are failing without it on OpFNegate for example).