Fix incorrect POD arg struct layout for kernels with long-vector POD args
Long-vector POD args are lowered to arrays. When these args are in a struct,
changing these member types can change the layout of the struct. This can
result in the clustered POD arg struct not having the same layout as the
corresponding reflection metadata.

Fix this behavior by using the same remapped type offset/size metadata as the
UBO Type Transform Pass to preserve the original layout.