add support for get image channel order and data_type (#882)
Add a pass to inline function containing either
function (get_channel_order or get_channel_data_type) to be able to
know which image is in input.

Add a pass to set metadatas on each kernel containing either function,
as well as metadata on each call to those functions.
Those metadatas are used in SPIRVProducer to generate to appropriate
offset and nonsemantic instructions.

Regroup RedeclareGlobalPushConstants under PushConstants.{cpp,h} to be
able to use them in SetImageChannelMetadataPass.

Do not use per-kernel push constant interfaces if the kernel call
either of the channel functions. We need the pod kernel arguments when
we redeclare globalpushconstant to add info about image channel.