fix alignment of structure when lowering long vectors
vectors are aligned on their size. But when lowering them to arrays,
it changes their alignment to be on the alignment of the element.

When lowering a structure make sure that we are keeping the offset
right. If not add a padding element.

Also make sure that the structure has the same size before and after
the lowering. If not add a padding element at the end.

As the lowered type might have more element than the original one,
rework the indices when lowering gep extractvalue or insertvalue