Implement printf
Add support for printf. The compiler will generate definitions that
store a unique identifier for the given format string and the value of
each argument.

A storage buffer of a configurable size is used to provide the printf buffer.
Information on the buffer and the individual printf calls is embedded in the
reflection instructions.

Also add a few misc fixes to make this work:
* Allow lowering of ConstantExprs assuming there is only 1 user

The below change has been pushed into a separate commit
due to rebase differences.
* Add support for atomic instruction implicit GEPs to ReplacePointerBitcastPass

Also, fixed printf_args.cl by replacing OpCompositeExtract with
OpUConvert and adjust the first component.

Also, fix with the usage of `PrintfBufferStorageBuffer` per reflection
specification, few the tests that were failing.

Also, test/printf_vector change the order of CHECK as needed.

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Signed-off-by: Pavan Lanka <pavan.l@samsung.com>