Stop trying to lower extremely long vectors.
Vector types wider than sixteen can be generated by optimizations, and if we
try to lower them we're prone to hitting asserts. The UndoInstCombine pass is
capable of handling long vectors in these cases so we're better off
leaving them for it to take care of.

The specific motivating case for this is if we include the libclc
conversion implementations in our builtins module and continue building
it with -O3 the resultant module has very long vectors in it that
UndoInstCombine handles better than LongVectorLowering does.