compiler: allow compilation of several inputs file together (#932)
* compiler: allow compilation of several inputs file together

This is a first step to fix OpenCL-CTS 'compiler' tests in clvk.

The idea is to use llvm IR as the intermediate language to link
program together.
To do that, this PR is now exposing one new function:
- CompileFromSourcesString: which will allow to take several input
files, compile and link them together into 1 output buffer.

Major changes:
- 'InputFilename' (now 'InputsFilename') allows for multiple inputs.
- New option 'IRBinaryOutput' force the LLVM IR to be dump in
binary. Either in a file, and/or in the output buffer.
- 'SetCompilerInstanceOptions' (and new function 'ProgramToModule')
expects 'program' to never be empty. The frontend 'Compile' function
is now making sure of it.
- Logs are now accumulated as many files can be parsed in one call to
clspv.

* Update lib/Compiler.cpp

Co-authored-by: Finlay <finlay.marno@codeplay.com>

* Update lib/Compiler.cpp

Co-authored-by: Finlay <finlay.marno@codeplay.com>

* Update lib/Compiler.cpp

Co-authored-by: Finlay <finlay.marno@codeplay.com>

* assert instead of returning error when appropriate

* unify output formats

* update old commet on emit-ir

* improve OutputFormat description

Co-authored-by: Finlay <finlay.marno@codeplay.com>