Fix definition of get_global_offset() when push constants are used (#1024)
GetPushConstantPointer is designed to return a GEP into the push constant
variable. When there is a single push constant or when indexing into the
first one (i.e. the global offset), this leads to a GEP with all zero
indices being created. With opaque pointers, this GEP is no longer created
and CreateInBoundsGEP returns the original pointer.
defineGlobalOffsetBuiltin was then assuming that the pointer returned
pointed to a specific push constant rather than the global push constant
variable itself which led to incorrect indices being used.

After this change, the extra indices are provided to GetPushConstantPointer
so the full GEP into the global offset vector is created directly.

It seems that this is a generic issue and users who want to guard against
this might now need to check whether CreateInBoundsGEP returned a
GEP instruction or the original pointer to guarantee that a chained
GEP uses the correct indices. This seems to break encapsulation.

Also support the case where the global push constant variable is
used by a load instruction. This happens in a clvk unit test.

Signed-off-by: Kévin Petit <kevin.petit@arm.com>