fix Rewrite Packed Structs pass for opaque pointers
For opaque pointers, do not look for packed structure in kernel
arguments.
Instead look for gep on packed structure or on structure with an
implicit cast.
Lowering structure (even unpacked) in implicit cast gep is needed to
avoid complicated issue (that I don't have a answer for) in
replacepointerbitcastpass.

Lower all those GEPs using 2 new functions to rework GEP indices from
a source type to a target type.
Also use those functions in Simplifypointerbitcastpass

SimplifyPointerBitcastPass needs to be updated with a new Map
containing ImplicitCasts (between 2 GEPs) that can be lower in this
pass. This is needed to avoid regression in the CTS.

Other minor fixes:
- ReplacePointerBitcastPass::ComputeLoad: CreateGEP might not create a
GEP, use getIndexedType to get the Load Type.
- BitcastUtils::getEleTypesBitWidths: fix typo in struct size
supported.

Fixes #1005
Fixes #1050

No regression on full cts with Swiftshader and Nvidia