Fix kernel arg info generation for unused arguments when opaque pointers are enabled
This is the pattern used by the CTS tests to validate clGetKernelArgInfo.

The information needed to report kernel arg info via the OpenCL API
is carried by ArgumentInfo reflection instructions. These are
made available via the last optional operand to Argument* reflection
instructions which means those need to be produced for unused arguments
too.

This changes relies on the metadata provided by Clang to deduce
the argument type. Pseudo-resources are allocated to descriptor set -1 /
binding -1 for unused arguments and runtimes need to deal
with this case (if the approach is accepted, this ought to be
documented as part of the specification for the ClspvReflection
instruction set).

This change also enables runtimes to correctly report errors when
kernel parameters are set to the wrong type of object (though it
forces the use of -cl-kernel-arg-info to do so).

Signed-off-by: Kévin Petit <kevin.petit@arm.com>