add LowerPrivatePointerPHIPass
This pass is trigger only if a PHI of a private pointer is detected.
Then, it will iterate through all the users of all allocas to only do
direct access.

GEP are removed, but we keep the offset added by it.
Load/Store are rework to add back the appropriate GEP before it.
PHI are rework to be about the offset and not the pointer itself.
Other instructions are not supported yet.

Fix #1071
Fix #723