improve clspv_get_transitive_libs performance
clspv_combine_static_lib is taking over 4min on my machine. This is
due to the recursive call to clspv_get_transitive_libs which is call
multiple times with the same arguments, and each time it is getting
properties that it already has to endup removing duplicates.

Instead, just create a variable to know that we have already run it
for one target, and do nothing if it has already been called.

This is reducing clspv_combine_static_lib to less that 1s on my
machine.