add support for vstore_half and generic pointers
add vstore_half and sincos to the builtins with generic pointer list.

Reorganise passes to have a ReplaceOpenCLBuiltinPass before and after
LowerAddrSpaceCastPass. Before is needed to deal with
to_global/to_private/to_local. After is needed to deal with
vstore_half.

Minor fixes in GetEquivalentBuiltinsWithoutGenericPointer.

Clean unused functions in InlineEntryPoints. This is needed to avoid
having dead instructions not removed in loweraddrspacecastpass.

Fixes in LowerAddrSpaceCastPass:
- fix LowerAddrSpaceCastPass since we updated when we run this pass
- remove function that have 1 use with is in fact coming from the
GlobalVariable used to make sure some libclc do not get removed
because not yet used.

Fix #1079