fix memcpy
As we are not capable of garante that a type won't change during the
compilation passes, we cannot use copymemory as types might have
changed between the time we create the node and when we want to
produce the equivalent spirv.
Thus always rely on load/store to perform the copy.

fix #1109