Implement Printf (#1117)
* Handle AtomicRMW instructions through implicit casts

The `OpAtomicIAdd` instructions need the pointer argument to be the same
type as result type. In the case of aggregate type, it needs to be
casted to base type, so use the implicit casts to find the containing
type.

(This change is inspired by the changes in PR#925)
Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Signed-off-by: Pavan Lanka <pavan.l@samsung.com>

* Implement printf

Add support for printf. The compiler will generate definitions that
store a unique identifier for the given format string and the value of
each argument.

A storage buffer of a configurable size is used to provide the printf buffer.
Information on the buffer and the individual printf calls is embedded in the
reflection instructions.

Also add a few misc fixes to make this work:
* Allow lowering of ConstantExprs assuming there is only 1 user

The below change has been pushed into a separate commit
due to rebase differences.
* Add support for atomic instruction implicit GEPs to ReplacePointerBitcastPass

Also, fixed printf_args.cl by replacing OpCompositeExtract with
OpUConvert and adjust the first component.

Also, fix with the usage of `PrintfBufferStorageBuffer` per reflection
specification, few the tests that were failing.

Also, test/printf_vector change the order of CHECK as needed.

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Signed-off-by: Pavan Lanka <pavan.l@samsung.com>

* Add documentation for printf support

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Signed-off-by: Pavan Lanka <pavan.l@samsung.com>

* Support printf with physical storage buffers

Signed-off-by: Callum Fare <callum@codeplay.com>

Fixed the formatting errors
Expect fail printf_physical_buffer.cl until we debug/root cause that

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Signed-off-by: Pavan Lanka <pavan.l@samsung.com>

* Add negative test for printf

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Signed-off-by: Pavan Lanka <pavan.l@samsung.com>

---------

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Signed-off-by: Pavan Lanka <pavan.l@samsung.com>
Co-authored-by: Gowtham Tammana <g.tammana@samsung.com>
Co-authored-by: Callum Fare <callum@codeplay.com>