fix physicalStorageBuffers when a kernel is calling another kernel
We used to leave the calling conv of callinst of spir_kernel function
as spir_kernel, while we had changed the function to be just a
spir_function.
This used to lead to instcombine just getting rid of the call as (I
think) it cannot find the function with this calling conv.

Set the calling conv to spir_func so that the call is not remove by
mistake.

This patch is fixing some compiler tests of the CTS when running with
physical addressing enabled.