fix printf when reaching end of printf buffer
If we do not have the space to store all the arguments, we should not
store them. It leads to memory corruption from out-of-bounds access on
platform like lavapipe.

I thought of merging the test for the printf id and the args, but we
need both of them.
Here is an example to understand:
Considering we have a printf with 1 argument and a buffer size of 8
bytes (2 words).
The first word is for the counter of number of bytes written. We are
left with one word.
When performing the printf, we will execute the atomic add, which will
set the counter to 2 (1 for the printfid and 1 for the arg).
If we do not store the printf id, the runtime won't be able to know
whether we are trying to print 2 string without argument (each one
needing 1 word) or if we are printing 1 string with 1 argument.