fix SmallerTypeNotAliasing when a structure is involved (#1167)
If the bigger type is a struct, we need to make sure that all its
fields are bigger or equal to the type selected to return. Otherwise,
return the smaller type amongst the field of the structure.

This fix ends up generating vector of 1 element. Fix getSPIRVConstant
and getSPIRVType to avoid generating vector of 1 element as it is not
allowed in SPIRV-V.

Fix #1166