gn: Add gn build infra
Setting up the infra for building clspv using GN[1]. At a high level the
build config mimics the LLVM build config but is independent in itself.
The GN build is setup to for building clspv both as a standalone and it
being a third_party in a top-level project. eg. ANGLE.

See `gn/README.md` for the usage.

[1]: https://gn.googlesource.com/gn/

Signed-off-by: Gowtham Tammana <g.tammana@samsung.com>
Change-Id: I1ea822eab9ab164cb076fa96150c580ba53d485f