fix to consider memcpy and memset with generic addrspace
When LowerAddrSpaceCastPass rework the function to remove the generic
addrspace part of it, it generates a name with a itanium mangling.

Take it into account in ReplaceLLVMInstrinsics so that those function
are taken care of.

Also add them to the DeadFunctions to avoid error later in the
verifier pass

Fix #1173