Implement read of 3D images with unnormalised sampler
We need to know where the sampler is coming from, so we need to inline
every function containing a read image of a 3d image with non literal
sampler.

This PR depends on KhronosGroup/SPIRV-Headers#377
We use the information from the sampler mask to know if you need to
use the original coordinates or the one we have normalised.

Like for channel_image_order and channel_image_data_type we use
metadata to pass information through the pipeline.