add hack to shuffle component of read_image
Some devices do not support VK_FORMAT_B8G8R8A8 with texel buffers.
As {CL_BGRA, CL_UNORM_INT8} is mandatory to be OpenCL compliant, those
devices have to find a map for it. A possible solution for those
devices is to map it to VK_FORMAT_R8G8B8A8. Using the swizzling we can
manage to have the proper order of component when reading using a
sampler, but for samplerless reads, we need to rework the component by
ourselves.