fix replacepointerbitcast for gep with result type smaller than src
If we cannot figure out the indices from constant offset using the
source type, it means that it is most probably smaller than the gep
result type. In that case use the gep result type as the source type.

Also if the source type is not a vector, array of struct, still
consider the scalar the go through it once.
We might need to merge it with some indice already existing.

Ref #1222