fix PointerBitcast considering type with null size
Types used to interface with SPIR-V have a null size (types like:
`{[0 x Ty]}`).
Try to avoid it by going through the type as long as SizeInBits
returns zero to get the real type size for it.

Do not do it for runOnUnneededCasts and runOnImplicitGEP because
either it is already taken into account in ReplacePointerBitcastPass,
or it can break the module to remove or insert gep.

Ref #1222