fix LowerPrivatePointerPHIPass by removing gep_map
When reusing geps, it is very complicated to make sure that the gep
dominates the instruction that will be using it.

I tried to force the gep to be just after the `Src` instruction, but
in that case, the gep might dominates some of its indices which is
wrong as well.

Let's just get rid of this optimisation idea for the time being.

Ref #1222