change infer type for phi nodes: consider incoming values
Refactor the part of InferType dealing with users type in a function.

For PHI Nodes, consider the type coming from the users, but also the
type from the incoming values.

The goal here is to end up with a reported type which is the bigger
one that all users and incoming values can easily bitcast to/from.

Previous implementation is leading to unsupported cases in pointer
bitcast passes for issue #1208 #1209 #1220 #1221 #1223